+++ src/GoogleBooksMatch.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * GoogleMatch class
+ * GoogleBooksMatch class
  *
  * @license    GPL 2 (http://www.gnu.org/licenses/gpl.html)
  * @author     mikespub
@@ -8,9 +8,10 @@
 
 namespace Marsender\EPubLoader;
 
-class GoogleMatch extends BaseMatch
+class GoogleBooksMatch extends BaseMatch
 {
     public const ENTITY_URL = 'https://www.googleapis.com/books/v1/volumes/';
+    public const ENTITY_PATTERN = '/^w+$/';
     public const CACHE_TYPES = ['google/authors', 'google/titles', 'google/volumes'];
     public const QUERY_URL = 'https://www.googleapis.com/books/v1/volumes?q={query}&maxResults={limit}&printType=books&projection=full&langRestrict={lang}';